CRUD bugs fixed

+++ service/user.service.ts
@@ -1,5 +1,6 @@
 import IUser from "../interface/user.interface"
 import userModel from "../models/user.model"
+import { ObjectId } from "mongodb"
 
 class UserService {
   private userModel = userModel
@@ -31,7 +32,8 @@ class UserService {
   }
 
   public async findOneUserService(userId: string) {
-    return await this.userModel.findOne({ userId })
+    console.log(userId)
+    return await this.userModel.findById(new ObjectId(userId))
   }
 
   public async updateUserService(
@@ -45,23 +47,26 @@ class UserService {
     posts?: [],
     savedPost?: []
   ) {
-    return await this.userModel.findByIdAndUpdate({
-      userId,
-      username,
-      password,
-      email,
-      followers,
-      following,
-      followRequest,
-      posts,
-      savedPost,
-    })
+    const newUser = await this.userModel.findByIdAndUpdate(
+      { _id: new ObjectId(userId) },
+      {
+        userId,
+        username,
+        password,
+        email,
+        followers,
+        following,
+        followRequest,
+        posts,
+        savedPost,
+      },
+      { new: true }
+    )
+    return newUser
   }
 
   public async deleteUserService(userId: string) {
-    return await this.userModel.findByIdAndDelete({
-      userId,
-    })
+    return await this.userModel.findByIdAndDelete({ _id: new ObjectId(userId) })
   }
 
   public async checkingUserService(username: string, password: string) {
